Change DOS_File flags from uint32_t to uint8_t
The flags are a bitfield specifically of 8 bits.
They get set by an 8 bit register inside DOS interrupts.
Nothing ever uses or sets anything wider than 8 bits.

+++ include/dos_system.h
@@ -129,7 +129,7 @@ public:
 
 	void SetDrive(uint8_t drv) { hdrive=drv;}
 	uint8_t GetDrive(void) { return hdrive;}
-	uint32_t flags   = 0;
+	uint8_t flags    = 0;
 	uint16_t time    = 0;
 	uint16_t date    = 0;
 	FatAttributeFlags attr = {};
