+++ src/dos/drive_overlay.cpp
@@ -215,8 +215,8 @@ public:
 			LOG_MSG("constructing OverlayFile: %s", name);
 	}
 
-	bool Write(uint8_t * data,uint16_t * size) override {
-		uint32_t f = flags&0xf;
+	bool Write(uint8_t* data, uint16_t* size) override {
+		uint8_t f = flags & 0xf;
 		if (!overlay_active && (f == OPEN_READWRITE || f == OPEN_WRITE)) {
 			if (logoverlay) LOG_MSG("write detected, switching file for %s",GetName());
 			if (*data == 0) {
